refactor: 더 직관적인 이름으로 수정
- socketMiddleware -> socketRouter
