+++ packages/backend/src/domains/player/inbound/player.controller.ts
@@ -1,7 +1,7 @@
 import { Socket } from "socket.io";
 import { PlayerPort } from "./player.port";
 import { PlayerService } from "../entity/palyer.service";
-import { SocketRouter } from "../../../utils/socketMiddlware";
+import { SocketRouter } from "../../../utils/socketRouter";
 import { Room } from "../../room/entity/room.entity";
 import { SearchRoomController } from "../../room/inbound/SearchRoom.api.controller";
